Fixed bug causing incorrect unwinding of layer files.
Layer files such as `esri/geometry`, `esri/renderer`, and `esri/symbol` contained a cache of all of their dependencies within a `require` call at the beginning of the file. This was causing the file to be reconstructed by `unwinder.js` incorrectly.

This new version removed the `require` cache at the beginning of the file and processes the rest of the file appropriately so that these modules can still be used.

Closes #18

Also added `grunt-node-inspector`.